Move data classes to model directory
