+++ app/src/main/java/com/kenchen/capstonenewsapp/NewsListViewHolder.kt
@@ -1,6 +1,7 @@
 package com.kenchen.capstonenewsapp
 
 import androidx.recyclerview.widget.RecyclerView
+import com.kenchen.capstonenewsapp.model.Article
 import com.kenchen.capstonenewsapp.views.ArticleView
 
 class NewsListViewHolder(private val articleView: ArticleView) : RecyclerView.ViewHolder(articleView) {
