+++ app/src/main/java/com/kenchen/capstonenewsapp/NewsService.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import android.util.Log
 import androidx.core.content.edit
 import androidx.preference.PreferenceManager
+import com.kenchen.capstonenewsapp.model.Article
 
 interface NewsService {
 }
